feat: expose resolved tool exposure reads
