+++ packages/contracts/src/index.ts
@@ -73,7 +73,7 @@ export { IosLocationGetCurrentArgs, IosCameraTarget, IosCameraCapturePhotoFormat
 // prettier-ignore
 export { AndroidLocationGetCurrentArgs, AndroidCameraTarget, AndroidCameraCapturePhotoFormat, AndroidCameraCapturePhotoArgs, AndroidAudioRecordClipArgs, AndroidActionArgs, AndroidLocationCoords, AndroidLocationGetCurrentResult, AndroidCameraCapturePhotoResult, AndroidAudioRecordClipResult, AndroidActionResult } from "./android.js";
 // prettier-ignore
-export { AgentAccessDefaultMode, AgentModelConfig, AgentSkillConfig, AgentMcpConfig, AgentToolConfig, AgentConversationConfig, AgentAttachmentInputMode, AgentAttachmentConfig, BuiltinMemoryServerSettings, AgentPersona, AgentSecretReference, AgentSecretReferences, AgentConfig, IdentityStyle, IdentityFrontmatter, IdentityPack, SkillRequires, SkillFrontmatter, SkillManifest, SkillProvenanceSource, SkillStatus, McpServerSpec, AgentTurnRequest, AgentTurnResponse, AgentListItem, AgentListResponse, AgentConfigListItem, AgentConfigListResponse, AgentConfigGetResponse, AgentConfigUpdateRequest, AgentConfigUpdateResponse, ManagedAgentSummary, ManagedAgentListResponse, ManagedAgentDetail, ManagedAgentGetResponse, ManagedAgentCreateRequest, ManagedAgentUpdateRequest, ManagedAgentRenameRequest, ManagedAgentRenameResponse, ManagedAgentDeleteResponse, AgentStatusResponse, AgentSkillCapability, AgentMcpCapabilitySource, AgentMcpCapability, AgentToolCapability, AgentCapabilitiesResponse } from "./agent.js";
+export { AgentAccessDefaultMode, AgentToolExposureBundle, AgentToolExposureTier, AgentToolExposureSelection, AgentToolExposureReadModel, AgentModelConfig, AgentSkillConfig, AgentMcpConfig, AgentToolConfig, AgentConversationConfig, AgentAttachmentInputMode, AgentAttachmentConfig, BuiltinMemoryServerSettings, AgentPersona, AgentSecretReference, AgentSecretReferences, AgentConfig, IdentityStyle, IdentityFrontmatter, IdentityPack, SkillRequires, SkillFrontmatter, SkillManifest, SkillProvenanceSource, SkillStatus, McpServerSpec, AgentTurnRequest, AgentTurnResponse, AgentListItem, AgentListResponse, AgentConfigListItem, AgentConfigListResponse, AgentConfigGetResponse, AgentConfigUpdateRequest, AgentConfigUpdateResponse, ManagedAgentSummary, ManagedAgentListResponse, ManagedAgentDetail, ManagedAgentGetResponse, ManagedAgentCreateRequest, ManagedAgentUpdateRequest, ManagedAgentRenameRequest, ManagedAgentRenameResponse, ManagedAgentDeleteResponse, AgentStatusResponse, AgentSkillCapability, AgentMcpCapabilitySource, AgentMcpCapability, AgentToolCapability, AgentCapabilitiesResponse } from "./agent.js";
 // prettier-ignore
 export { ExplicitDedicatedToolId, RoutedToolTargeting, RoutedToolSelectionMode, RoutedToolSelectedNode, RoutedToolExecutionMetadata, SecretReferenceId, SecretReferenceAlias, SecretReferenceSelector, SecretCopyToNodeClipboardArgs } from "./routed-tool.js";
 export {
