+++ packages/gateway/src/app/modules/agent/tool-exposure-read-model.ts
@@ -1,0 +1,1 @@
+export { resolveAgentToolExposureReadModel } from "../../../modules/agent/tool-exposure-read-model.js";
