+++ packages/operator-ui/tests/operator-ui.a11y.shared.ts
@@ -259,6 +259,10 @@ function sampleAgentStatusResponse() {
       },
     ],
     tools: ["shell", "http"],
+    tool_exposure: {
+      mcp: {},
+      tools: {},
+    },
     conversations: {
       ttl_days: 365,
       max_turns: 0,
