+++ packages/operator-ui/tests/pages/agents-page-editor.test-helpers.tsx
@@ -33,6 +33,10 @@ export function sampleManagedAgentDetail(agentKey: string) {
         character: "architect",
       },
     }),
+    tool_exposure: {
+      mcp: {},
+      tools: {},
+    },
     identity: IdentityPack.parse({
       meta: {
         name: agentKey === "default" ? "Default Agent" : "Agent One",
