+++ packages/transport-sdk/tests/http-client.test-ops-support.ts
@@ -1,7 +1,9 @@
+import { registerHttpClientAgentExposureTests } from "./http-client.test-agent-exposure-support.js";
 import { registerHttpClientOpsAdminTests } from "./http-client.test-ops-admin-support.js";
 import { registerHttpClientOpsCoreTests } from "./http-client.test-ops-core-support.js";
 
 export function registerHttpClientOpsTests(): void {
+  registerHttpClientAgentExposureTests();
   registerHttpClientOpsCoreTests();
   registerHttpClientOpsAdminTests();
 }
